+++ lib/src/grinder.dart
@@ -1,8 +1,6 @@
 // Copyright 2015 Google. All rights reserved. Use of this source code is
 // governed by a BSD-style license that can be found in the LICENSE file.
 
-library grinder.src.grinder;
-
 import 'dart:async';
 
 import 'package:cli_util/cli_logging.dart' show Ansi;
@@ -27,7 +25,7 @@ set defaultTask(GrinderTask? v) {
 }
 
 /// A class representing a running instance of a Grinder.
-class Grinder {
+final class Grinder {
   final List<GrinderTask> _tasks = [];
   Map<GrinderTask, List<TaskInvocation>>? _taskDeps;
   final List<TaskInvocation> _invocationOrder = [];