+++ lib/src/grinder_exception.dart
@@ -1,10 +1,8 @@
 // Copyright 2015 Google. All rights reserved. Use of this source code is
 // governed by a BSD-style license that can be found in the LICENSE file.
 
-library grinder.src.grinder_exception;
-
 /// An exception class for the Grinder library.
-class GrinderException implements Exception {
+final class GrinderException implements Exception {
   /// A message describing the error.
   final String message;