+++ lib/src/grinder_task.dart
@@ -1,8 +1,6 @@
 // Copyright 2015 Google. All rights reserved. Use of this source code is
 // governed by a BSD-style license that can be found in the LICENSE file.
 
-library grinder.src.grinder_task;
-
 import 'dart:collection';
 
 import 'grinder_context.dart';
@@ -15,7 +13,7 @@ typedef TaskFunction = dynamic Function(TaskArgs args);
 
 /// Represents a Grinder task. These can be created automatically using the
 /// [Task] and [Depends] annotations.
-class GrinderTask {
+final class GrinderTask {
   /// The name of the task.
   final String name;