+++ lib/src/run.dart
@@ -1,8 +1,6 @@
 // Copyright 2015 Google. All rights reserved. Use of this source code is
 // governed by a BSD-style license that can be found in the LICENSE file.
 
-library grinder.src.run;
-
 import 'dart:async';
 import 'dart:convert';
 import 'dart:io';
@@ -132,7 +130,7 @@ Future<String> runProcessAsync(String executable,
 }
 
 /// An exception from a process which exited with a non-zero exit code.
-class ProcessException implements Exception {
+final class ProcessException implements Exception {
   final String executable;
   final int exitCode;
   final String stdout;
@@ -146,7 +144,7 @@ class ProcessException implements Exception {
 
 /// Arguments passed to [Process.run] or [Process.start].
 /// See [Process.run] for more details.
-class RunOptions {
+final class RunOptions {
   final String? workingDirectory;
   final Map<String, String> environment;
   final bool includeParentEnvironment;