+++ lib/src/singleton.dart
@@ -1,8 +1,6 @@
 // Copyright 2015 Google. All rights reserved. Use of this source code is
 // governed by a BSD-style license that can be found in the LICENSE file.
 
-library grinder.src.singleton;
-
 import 'dart:io';
 
 import '../grinder.dart';