+++ lib/src/task_invocation.dart
@@ -1,13 +1,11 @@
 // Copyright 2015 Google. All rights reserved. Use of this source code is
 // governed by a BSD-style license that can be found in the LICENSE file.
 
-library grinder.src.task_invocation;
-
 /// An invocation of a [GrinderTask].
 ///
 /// Identifies the [name] of the task to invoke, and any [arguments] to send to
 /// it.
-class TaskInvocation {
+final class TaskInvocation {
   final String name;
   final TaskArgs _arguments;
 
@@ -37,7 +35,7 @@ class TaskInvocation {
 ///
 /// The arguments could be in the form of flags (`grind foo --release`) or flags
 /// (`grind foo --config=bar`).
-class TaskArgs {
+final class TaskArgs {
   final String taskName;
 
   /// The original list of arguments that were parsed.