+++ test/grinder_test.dart
@@ -1,14 +1,12 @@
 // Copyright 2013 Google. All rights reserved. Use of this source code is
 // governed by a BSD-style license that can be found in the LICENSE file.
 
-library grinder_test;
-
 import 'dart:async';
 
 import 'package:grinder/grinder.dart' hide fail;
 import 'package:test/test.dart';
 
-import 'src/_common.dart';
+import 'src/common.dart';
 
 void main() {
   group('grinder', () {