+++ test/integration_test.dart
@@ -1,13 +1,11 @@
 // Copyright 2014 Google. All rights reserved. Use of this source code is
 // governed by a BSD-style license that can be found in the LICENSE file.
 
-library grinder.integration_test;
-
 import 'package:grinder/grinder.dart';
 import 'package:grinder/src/cli.dart';
 import 'package:test/test.dart';
 
-import 'src/_common.dart';
+import 'src/common.dart';
 
 Map ranTasks = {};