+++ test/src/cli_util_test.dart
@@ -1,15 +1,13 @@
 // Copyright 2015 Google. All rights reserved. Use of this source code is
 // governed by a BSD-style license that can be found in the LICENSE file.
 
-library grinder.src.cli_test;
-
 import 'package:grinder/src/cli.dart';
 import 'package:grinder/src/cli_util.dart';
 import 'package:grinder/src/grinder.dart';
 import 'package:grinder/src/grinder_task.dart';
 import 'package:test/test.dart';
 
-import '_common.dart';
+import 'common.dart';
 
 void main() {
   group('cli_util', () {