+++ test/src/common.dart
@@ -1,8 +1,6 @@
 // Copyright 2014 Google. All rights reserved. Use of this source code is
 // governed by a BSD-style license that can be found in the LICENSE file.
 
-library grinder.test.common;
-
 import 'dart:async';
 
 import 'package:grinder/grinder.dart';
@@ -20,7 +18,7 @@ void grinderTest(String name, Function setup, TestVerification verify) {
   });
 }
 
-class MockGrinderContext implements GrinderContext {
+final class MockGrinderContext implements GrinderContext {
   @override
   Grinder get grinder =>
       throw UnsupportedError('MockGrinderContext.grinder is unsupported');