+++ test/src/task_discovery/bad_tasks.dart
@@ -3,8 +3,6 @@
 
 // ignore_for_file: prefer_void_to_null
 
-library grinder.test.task_discovery.bad_tasks;
-
 import 'package:grinder/grinder.dart';
 
 import 'external_tasks.dart';
@@ -18,7 +16,7 @@ const dynamic nullTask = null;
 Null get nullReturningGetter => null;
 
 @Task()
-class Class {}
+final class Class {}
 
 @Task()
 @Depends(shownMethod)