+++ test/src/task_discovery/external_tasks.dart
@@ -1,8 +1,6 @@
 // Copyright 2015 Google. All rights reserved. Use of this source code is
 // governed by a BSD-style license that can be found in the LICENSE file.
 
-library grinder.test.task_discovery.external_tasks;
-
 import 'package:grinder/grinder.dart';
 
 @Task('foo description')