+++ test/src/task_invocation_test.dart
@@ -1,8 +1,6 @@
 // Copyright 2013 Google. All rights reserved. Use of this source code is
 // governed by a BSD-style license that can be found in the LICENSE file.
 
-library grinder.src.task_invocation_test;
-
 import 'package:grinder/grinder.dart' hide fail;
 import 'package:test/test.dart';